Revert "Merge branch 'main' of https://github.com/VERSO-UVM/vrwa-trainingdb into course-manager"
This reverts commit d547b39c1388be3f6ca9b746aabbb19fc04ba65e, reversing
changes made to d409be971716813c3931ddd91bfa8503c92bff33.
