+++ backend/src/trpc.ts
@@ -1,17 +1,13 @@
 import { adminRouter } from "~/routers/admin";
 import { authRouter } from "~/routers/auth";
 import { certificateRouter } from "~/routers/certificate";
-import { reservationRouter } from "~/routers/reservation";
-import { router } from "~/utils/trpc";
 import { courseManagerRouter } from "./routers/courseManager";
-import { profileRouter } from "./routers/profile";
+import { router } from "./utils/trpc";
 
 export const appRouter = router({
   authRouter,
   adminRouter,
   certificateRouter,
-  reservation: reservationRouter,
-  profile: profileRouter,
   courseManagerRouter,
 });
