--- webapp/app/admin/instructor-manager.tsx
@@ -1,13 +1,0 @@
-import { PageHeader } from "~/components/page-header";
-
-export function InstructorManager() {
-  return (
-    <>
-      <PageHeader>Instructors</PageHeader>I think we might need a separate page
-      to manage instructors. OR alternatively we can make the trainee page a
-      general user management page. It's just the main workflows are *manually
-      changing trainee credit hours* and *manually changing trainee attendance*.
-      -R
-    </>
-  );
-}
