+++ webapp/app/admin/navigation.tsx
@@ -46,14 +46,6 @@ export function AdminSidebar() {
             </SidebarMenuItem>
             <SidebarMenuItem>
               <SidebarMenuButton asChild>
-                <Link to="/admin/instructors">
-                  <Users />
-                  Instructors
-                </Link>
-              </SidebarMenuButton>
-            </SidebarMenuItem>
-            <SidebarMenuItem>
-              <SidebarMenuButton asChild>
                 <Link to="/admin/certifications">
                   <Trophy />
                   Certifications
