+++ webapp/app/components/ui/data-table-header.tsx
@@ -1,7 +1,7 @@
 import { flexRender, type Table } from "@tanstack/react-table";
+import { TableHead, TableHeader, TableRow } from "./table";
+import { Button } from "./button";
 import { ArrowDownZA, ArrowUpAZ, ArrowUpDown } from "lucide-react";
-import { TableHead, TableHeader, TableRow } from "~/components/ui/table";
-import { Button } from "~/components/ui/button";
 
 interface DataTableHeaderProps<TData> {
   table: Table<TData>;
@@ -9,8 +9,7 @@ interface DataTableHeaderProps<TData> {
 
 export function DataTableHeader<TData>({ table }: DataTableHeaderProps<TData>) {
   return (
-    // Fix header at top while scrolling!
-    <TableHeader className="sticky bg-background top-0 z-10">
+    <TableHeader>
       {table.getHeaderGroups().map((headerGroup) => (
         <TableRow key={headerGroup.id}>
           {headerGroup.headers.map((header) => {
