+++ webapp/app/components/ui/table.tsx
@@ -6,7 +6,7 @@ function Table({ className, ...props }: React.ComponentProps<"table">) {
   return (
     <div
       data-slot="table-container"
-      className="relative w-full overflow-x-auto max-h-150 overflow-y-auto"
+      className="relative w-full overflow-x-auto"
     >
       <table
         data-slot="table"
