+++ webapp/app/root.tsx
@@ -94,8 +94,7 @@ export function Layout({ children }: { children: React.ReactNode }) {
           </head>
           <body>
             <TooltipProvider>{children}</TooltipProvider>
-            {/* customize getKey to prevent scroll reset when URLSearchParams changes*/}
-            <ScrollRestoration getKey={(location) => location.pathname} />
+            <ScrollRestoration />
             <Scripts />
           </body>
         </html>
