--- webapp/app/routes/admin/instructors.tsx
@@ -1,10 +1,0 @@
-import { InstructorManager } from "~/admin/instructor-manager";
-import type { Route } from "../+types/home";
-
-export function meta({}: Route.MetaArgs) {
-  return [{ title: "Admin - Instructor Manager" }];
-}
-
-export default function () {
-  return <InstructorManager />;
-}
