+++ webapp/app/routes/admin/trainees.tsx
@@ -1,5 +1,5 @@
 import type { Route } from "../+types/home";
-import { TraineeManager } from "~/admin/trainee-manager";
+import { TraineeManager } from "~/admin/trainees";
 
 export function meta({}: Route.MetaArgs) {
   return [{ title: "Admin - Trainee Manager" }];
