+++ webapp/package.json
@@ -21,7 +21,6 @@
     "@trpc/tanstack-react-query": "^11.10.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "drizzle-zod": "^0.8.3",
     "isbot": "^5.1.35",
     "lucide-react": "^0.575.0",
     "radix-ui": "^1.4.3",
